add two-digit-x-two-digit dictionary
