+++ src/app/[lang]/two-digit-x-two-digit/QuestionBox.tsx
@@ -1,0 +1,35 @@
+'use client'
+
+import { QuestionArea } from '@/app/[lang]/two-digit-x-two-digit/QuestionArea'
+import { useState } from 'react'
+
+import questionAreaStyle from '@/app/[lang]/two-digit-x-two-digit/questionArea.module.scss'
+import { DICTIONARY } from '@/i18n/dictionaries'
+
+type QuestionBoxProps = {
+    dict: DICTIONARY
+}
+
+export function QuestionBox(questionBoxProps: QuestionBoxProps) {
+    const { dict } = questionBoxProps
+    const [questionCount, setQuestionCount] = useState<number>(1)
+
+    const clickHandlerNext = () => {
+        setQuestionCount((count) => count + 1)
+    }
+
+    return (
+        <>
+            <div className={questionAreaStyle.container}>
+                <div className={questionAreaStyle.question}>
+                    Q.{questionCount}
+                </div>
+                <QuestionArea
+                    key={questionCount}
+                    clickHandlerNext={clickHandlerNext}
+                    dict={dict}
+                />
+            </div>
+        </>
+    )
+}
