+++ src/app/[lang]/two-digit-x-two-digit/questionArea.module.scss
