+++ src/i18n/locale/en-US/two-digit-x-two-digit.json
@@ -1,0 +1,4 @@
+{
+    "next": "Next",
+    "answer": "Answer"
+}
