+++ src/i18n/locale/fr/two-digit-x-two-digit.json
@@ -1,0 +1,4 @@
+{
+    "next": "Sur suivant",
+    "answer": "Réponse"
+}
