+++ src/i18n/locale/ja/two-digit-x-two-digit.json
@@ -1,0 +1,4 @@
+{
+    "next": "つぎへ",
+    "answer": "こたえ"
+}
